requirements.txt
streamlit==1.22.0
torch>=2.0.0
transformers==4.28.1
spacy==3.5.2
scikit-learn==1.2.2
pandas==2.0.0
pyyaml==6.0
pillow==10.1.0